backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
httpx==0.25.1
pydantic==2.11.0
redis==5.0.1
python-dotenv==1.0.0
google-genai==0.1.0
